Further work on Recipes Pages
Clickable Recipes!

When clicked, you will be taken to a page which will have a summary, instructions, and ingredients buttons that can take you to different breakdowns of the recipe.

Searchable Recipes!

On the Recipe Database page, there is now a search bar at the top that allows you to search for a specific Recipe.

Known issues :(

Formatting will need work, as currently the page moves around a lot when moving between tabs.

Page Routing will also need adjustments, as you are unable to go back to the database from the recipe page.